Pull Request to add my work for Orca-Python AYRNA's repository  (#2)
* G&S Added

Getters & Setters of the NNPOM class variables (epsilonInit, hiddenN, iter and lambdaValue), added.

* Method: "__convertThresholds" added 

Private method of class NNPOM, "__convertThresholds" added. This method transforms thresholds to perform unconstrained optimization.

* Method: "__unpackParameters" added 

Private method of class NNPOM, "__convertThresholds" added. This method gets Theta1, Theta2 and thresholds_param back from the whole array nn_params.

* Method: "__randInitializeWeights" added 

Private method of class NNPOM, "__randInitializeWeights" added. This method randomly initializes the weights of a layer with L_in incoming connections and L_out outgoing connections.

* File "nnpom_test.json" added

Json test file for the NNPOM class created with the initial values ​​for the class variables initialized by default.

* Identation errors fixed

Indentation errors caused by the IDE fixed (transformation of tabs into spaces, mainly)

* NNPOM.py : Method "__nnPOMCostFunction" added

Private method of class NNPOM, "__nnPOMCostFunction" added. This method implements the cost function and obtains the corresponding derivatives.

* Little errors fixed on "__nnPOMCostFunction"

* Fixed some bugs into "__nnPOMCostFunction" method

* NNPOM.py: Method "fit" almost done

Almost done method except the call to the back-propagation error method.

* NNPOM.py: Method "predict" done

Method "predict", belonging to the NNPOM class done.

* NNPOM.py: Method "fitpredict" erase and some GyS added

- Method "fitpredict" erase.
- New Getters and Setters added for new variables (theta1, theta2, thresholds, num_labels and m).

* Datasets from NNPOM class added

* nnpom_test.json updated

* test_nnpom.py updated

* NNPOM.py some bugs fixed

* NNOP.py: NNOP class created

* NNPOM.py: some bugs fixed 2

Some comments added and others fixed. Also a few lines of code fixed.

* NNPOM.py: some bugs fixed 3

Some comments added and others fixed. Also a few lines of code fixed.

* NNOP.py: NNOP class almost finished

All the methods of the NNOP class have been created and coded.

* nnop_test.json created

nnop_test.json file created

* test_nnop.py created

test_nnop.py file created

* Rename some files and change a few parameters of NNPOM executing config

Some modifications have been implemented in the name of some files and in the execution configuration of the NNPOM algorithm to test that the algorithm works optimally

* A error in NNPOM.py fixed and changed a config of utilities.py

A error in NNPOM.py fixed and changed a config of utilities.py to prove the algorithm

* Changes to the NNPOM algorithm configuration file

The line of the configuration file that managed the input preprocessing has been eliminated, so that there is no preprocessing and test that the data obtained is identical with respect to the equivalent algorithm of the ORCA project.

The config.py file has also been modified, which was modified so that there was no preprocessing and leave it as it was at the beginning

* predict function updated

Method predict of NNPOM class updated with some bugs fixed

* Trying some configurations. Convergence of NNPOM not obtained yet.

* Some bugs fixed

Optimize function changed and some error of sintaxis fixed too

* NNPOM: Code cleared and comments fixed

Algorithm finished. It would only be necessary to carry out the test files and add the databases to carry out the experiments

* Revert "Algorithm mostly finished"

This reverts commit ce6c5023d415d639c0a8d451ea4c7bd2dc5405ee.

* NNPOM Algorithm finished

* NNPOM experiments done

* NNPOM Clean code

* NNOP: Advance with the algorithm

* NNOP: It seems that NNOP Algorithm converges.

* NNOP. An experiment was realized

* Integration test in progress

* Integration tests passed!

* Name of Tests classes updated